ID3: Fix v2.4 extended header handling - Test invalid extended header
Add tests to read v2.4 files with invalid extended header.

The test file 'v2.4-ext-header-invalid-too-short.mp3' is a copy of test
file 'v2.4-ext-header.mp3', but with the extended header size edited to
become 3. This too short, it must be a minimum of 4 bytes.

The test file 'v2.4-ext-header-invalid.mp3' is a copy of test file
'v2.4-ext-header.mp3', but with the extended header size edited to
become 1 more than the size of the entire id3 tag. This is too long.